multiple admin

+++ src/Components/AdminDashBoard/ViewExam.jsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import '../../Style/AdminPagesStyle/StyleStudentView.css';
+import "../../Style/AdminPagesStyle/StyleViewExam.css"
 import axios from 'axios';
+import copy from 'clipboard-copy';
 import refreshicon from '../../ImagesAndLogo/refresh.png';
-
-function ViewExam() {
+import gif from "../../ImagesAndLogo/loading-loading-forever.gif";
+function ViewExam({adminEmail,adminId}) {
   const [examData, setExamData] = useState([]);
-
+const [result,setResult]=useState(true);
   useEffect(() => {
     handleExamData(); // Fetch data when the component mounts
   }, []);
@@ -16,9 +17,9 @@ function ViewExam() {
 
   const handleExamData = async () => {
     try {
-      const response = await axios.get('http://localhost:8080/api/getAllExams',config);
+      const response = await axios.get(`http://localhost:8080/api/getAllExams/${adminEmail}`,config);
+      setResult(false);
       setExamData(response.data);
-      console.log(response.data);
     } catch (e) {
       console.log(e);
     }
@@ -28,7 +29,7 @@ function ViewExam() {
     try {
       const confirmDelete = window.confirm('Are you sure you want to delete all Exam Data?');
       if (confirmDelete) {
-        await axios.delete('http://localhost:8080/api/deleteAllExams',config);
+        await axios.delete(`http://localhost:8080/api/deleteAllExams/${adminEmail}`,config);
         console.log('All exams deleted successfully from the database!');
         handleExamData(); // Refresh exam data after successful deletion
       } else {
@@ -59,19 +60,32 @@ function ViewExam() {
     return formattedDate;
   };
 
+  const handleCopyUrl = (url) => {
+    copy(url);
+    alert('URL copied to clipboard!');
+  };
+
   return (
     <>
-      <div>
-        <div className='AddStuTitle'>View Exam Data:</div>
+      <div className='viewExam-container'>
+        <div className='AddStuTitle'>Exams:</div>
         <br />
-        <button type="button" class="btn btn-outline-secondary" style={{ margin: "10px" }} onClick={handleExamData}>
+        <button type="button" class="btn btn-primary" style={{ margin: "10px" }} onClick={handleExamData}>
           Refresh {<img src={refreshicon} className='imgref' alt="refresh" />}
         </button>
 
         <br />
         <div>  <h5> <b>  Total Number Of Exams: {examData.length}  </b> </h5>  </div>
+
+        {result && (<div className='loading'>
+          <div className='innerOfLoading'>
+ Please wait, We are loading Exams...... {<img src={gif} className='gif' alt="refresh" />}
+         </div>
+         </div> )}
+        {!result && examData.length<=0 &&(<div className='no-exam'>There is no Exam, please create one.</div> )}
+        
         <div  className='tableDiv' >
-          <table class="custom-table">
+          <table class="table table-striped custom-table table-examView">
             <thead>
               <tr>
                 <th>Exam Name</th>
@@ -89,7 +103,12 @@ function ViewExam() {
                   <td>{formatTimestamp(exam.examSchedule)}</td>
                   <td>{exam.examDuration}</td>
                   <td>{exam.examId}</td>
-                  <td> http://localhost:3000/studentLogin/{exam.examId}</td>
+                  <td> http://localhost:3000/studentLogin/{exam.examId}/{adminId}
+                  <button
+                type="button"
+                className="btn btn-primary btn-copy btn2 "
+                onClick={() => handleCopyUrl(`http://localhost:3000/studentLogin/${exam.examId}/${adminId}`)}
+              >copy</button> </td>
                   <td>
                     <button type="button" class="btn btn-danger" onClick={() => handleDeleteById(exam.examId)}>
                       Delete
